stratgen/stratgen/requirements.txt
# --- Core API & Utils ---
fastapi>=0.110
uvicorn[standard]>=0.29
pydantic>=2.6
httpx>=0.27
pyyaml>=6.0
tenacity>=8.2

# --- RAG / Retrieval ---
llama-index>=0.10
qdrant-client>=1.7
sentence-transformers>=2.6
FlagEmbedding>=1.2.10
tiktoken>=0.6

# --- Parsing / Ingestion ---
unstructured[pdf]>=0.14
pymupdf>=1.23

# --- PowerPoint / Rendering ---
python-pptx>=0.6.23
jinja2>=3.1
pillow>=10
matplotlib>=3.8
